refactor(InputView): 숫자 검증 로직을 BridgeSize로 위임
숫자 검증 로직을 InputView에서 BridgeSize 클래스로 위임하였습니다. 다리의 길이가 숫자인지 아닌지는 객체가 판단해야 한다고 생각했기 때문입니다.
