+++ src/test/java/bridge/domain/validation/InputValidatorTest.java
@@ -1,6 +1,5 @@
 package bridge.domain.validation;
 
-import static bridge.domain.validation.InputValidator.INPUT_ONLY_NUMBERS;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import org.junit.jupiter.params.ParameterizedTest;
@@ -8,11 +7,5 @@ import org.junit.jupiter.params.provider.ValueSource;
 
 class InputValidatorTest {
 
-    @ParameterizedTest(name = "숫자가 아닌 값을 입력하면 예외가 발생한다. 입력: {0}")
-    @ValueSource(strings = {"하하", "hoho", " ", "1234a"})
-    void validateThisIsNumber(String inputValue) {
-        assertThatThrownBy(() -> InputValidator.validateThisIsNumber(inputValue))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessageContaining(INPUT_ONLY_NUMBERS);
-    }
+
 }
